Failing tests and a hacky fix for the namespacing/inverse_foreign_key issue
When declaring `references_and_referenced_in_many` relationships, the IDs for related objects are stored in a document using a key based on the association name, rather than the class name. For example:

  module Blah
    class Thing
      include Mongoid::Document
      references_and_referenced_in_many :widgets, :class_name => "Blah::Widget"
    end
    class Widget
      include Mongoid::Document
      references_and_referenced_in_many :things, :class_name => "Blah::Thing"
    end
  end

  thing = Blah::Thing.create
  widget = thing.widgets.create

The object `thing` will now have a property called `widget_ids`, which it uses to load its associated `Blah::Widget` instances.

The problem is when you try to use criteria to query this associations. Filtering by name, for example:

  thing.widgets.where(:name => "alpha")

will try to query using 'blah/widget_ids' as the foreign key, whereas the ids are actually stored in `widget_ids`.

This fix is very hacky, but I'm not yet familiar enough with the internals of Mongoid to do anything more elegant. At the very least, nothing else seems to be affected.