Preserve milliseconds when persisting Time and DateTime.
This time stuff is a mess. I'm not happy with my solution.
Although it seems to work. One of the things that caused me trouble
was Time#usec which should return the microseconds but seem to return
the milliseconds:

t = Time.now        #=> 2007-11-19 08:03:26 -0600
"%10.6f" % t.to_f   #=> "1195481006.775195"
t.usec              #=> 775195

Thatswhy it is not usable in the examples.

Also I tried to archive the same behavior requested for
Time in issue #64 for DateTime because its really the same.

Any Feedback is really appreciated.