Set many to many foreign key defaults lazily.
This avoids persisting empty arrays if the relation has not been
accessed. Once the relation is accessed in any manner, even if nothing
is set, empty arrays can be persisted on insert.

If the document is being updated, we guard against persisting empty
arrays only if they were previously empty or nil.

[ close #1564 ]