Don't apply options on sync callback updates.
When a default_scope exists on either side of a many-to-many relation
that applies options (like a sort), the resulting MongoDB command to
update the keys is invalid since MongoDB doesn't like $and queries
inside an update with a $query and $orderby options. It raises no error
in safe mode, but does log the following error in the db:

exception: invalid operator: $and code:10068 0ms

This fix removes any options in the criteria when updating the keys.

[ fix #2280 ]