Fix removal on embedded document via nested attrs.
Previously, embedded documents would get removed from the relation on a
nested attributes delete immediately, which would cause issues if the
parent document validation failed. Now these removals are delayed until
the parent passes validation and goes to persist.

Note this is a behaviour change, but since the API does not change and
the previous behaviour was incorrect and did not match AR this was able
to go into a point release. But it is worth nothing for those expecting
the old behaviour and hacking around it - they can remove the
workarounds at this point.

[ fix #2367 ]