Add short versions of timestamps. (Rodrigo Saito)
This adds the following modules that corresponds to the normal timestamp
modules, with the exception that the fields are stored in the database
as c_at and u_at to save space.

Mongoid::Timestamps::Short - for both created and updated.
Mongoid::Timestamps::Created::Short - for only a created.
Mongoid::Timestamps::Updated::Short - for only an updated.

[ close #2525 ]
[ close #2419 ]