Don't wrap a class that behaves as an array inside another array
Seeing an issue where attempting an autosave on a has_many fails because relation is a Mongoid::Relations::Targets::Enumerable. When array attempts to wrap this, it results in a nested array rather than a single array as expected. Adding a check for kind_of? Array avoids wrapping a deep class that already behaves as an array.