Fixed bug in relationships with `inverse_of` field
This fixes a bug in `Mongoid::Relations::Referenced::Many` when it goes
to find the inverse of a relationship within the
`#with_inverse_field_criterion` method. The method was trying to use
`:inverse_of_field` which doesn't exist on the metadata and is, instead,
`:inverse_of', resulting in a method-not-found error.