Remove Attributes::Readonly#loaded? method
Before this change, Attributes::ReadOnly defined a private `#loaded?`
method on every mongoid document.

Since it is a common name for a public method, this may be seen as
namespace pollution. In order to reduce the pollution, this method was
dissolved in its one and only call and is removed with this change.