Make {attr}_before_type_cast behave same as read_attribute_before_type_cast
Prior to this commit, #{attr}_before_type_cast() was falling back to
send(attr), whereas read_attribute_before_type_cast() was falling back
to read_raw_attribute(). In most cases this worked, but if a getter for
the attribute was in place and no pre-typecasted value was available,
they could return different values.

This remedies the issue by causing #{attr}_before_type_cast() to simply
be a shortcut to read_attribute_before_type_cast(attr).