Add in setter method check for attribute before creating new
We're using `#clone` in production, and have had to occasionally remove fields from classes, and did not see a need to fully migrate the data to unset the attribute on each instance. 

However, we're now running into `NoMethoError`s when calling `.clone` because it uses `self.class.new(attrs)` to instantiate the a data-filled object. I guess an alternative might be to literally call `instantiate` on the class and then tap into it, cycle through attrs in the same `dynamic_attrs` is assigned. I guess it comes down to if we want it to be a proper clone, with all of the attributes, even the ones without getters/setters on the class presently, or just a clone of a valid object based on the fields that are currently configured on the class.

Any thoughts?