Fix nil preventing fallbacks on localized fields
In 8fef344a1a828a46e6cd58e73faa6d5d57afe0ed, a fix to prevent falling
back to another locale when the current locale's value was `false`
introduced a regression where the value `nil`, which was previously
treated as a missing translation, became accepted as a translated value
too.

This PR assumes that was an unwanted change, fixes it, and introduces a
regression test.