ActionController::Parameters are not properly sanitized when instantiating embedded/referenced relations and localized fields.
Fortunately the current behavior does NOT cause a security issue because an MethodMissing error happens in all cases (i.e. Mongoid chokes on processing the ActionController::Parameters when it expects a Hash)

This PR standardizes and tests the behavior.

In addition, I am converting the current single usage of #sanitize_for_mass_assignment to #sanitize_forbidden_attributes. Although the two are aliased in ActiveModel, ActiveRecord uses the latter only and never the former. There are some gems, for example ProtectedAttributesContinued that work with the latter only.