Dead functionality removal: Mongoid::Association::Referenced::NestedAttributes::One
The class Mongoid::Association::Referenced::NestedAttributes::One is not used anywhere in the code, and moreover, attempting to load it will result in an error because its supposed parent class "NestedBuilder" does not exist. It looks like this class was forgotten to be removed in a refactor done ages ago. Does not need changelog entry.