Add pretty print support (#5810)
* Add pretty print support

This adds the `pretty_print` method, which pretty prints the same information as the `inspect` method. This is meant to be called by the default 'pp' gem, e.g. by calling `pp(person)` or `person.pretty_inspect`.

The specs are an almost identical to the specs for the `inspect` method to ensure the same information is included.

* Simplify code with if modifier

Co-authored-by: Jamis Buck <jamisbuck@gmail.com>

* Avoid "default gem" terminology

Co-authored-by: Jamis Buck <jamisbuck@gmail.com>

---------

Co-authored-by: Jamis Buck <jamisbuck@gmail.com>
Co-authored-by: Jamis Buck <jamis.buck@mongodb.com>