MONGOID-5839 Fix eager-loading from STI subclasses (backport to 8.0)
(backport to 8.0)

If the root of a query is an STI subclass (e.g. Subclass.all) AND
the query tries to eager load (includes) another association, the
eager load was failing because it was looking for inverse
associations using the STI subclass name, instead of the class at
the root of the hierarchy.