MONGOID-5872 - [🛠️ Let's Fix Mongoid's Broken CI] Actually test Rails 6.0.0, 7.0.0, 7.1.0, and released version of 8.0.0 (#5980)
* Actually test Rails 6.0.0, 7.0.0, 7.1.0, and released version of 8.0.0.

* Update rails-7.0.gemfile

We have MONGOID-5812 for separately addressing Ruby 3.4 support

* Skip 6.0 for the BSON::Decimal128 validation

Rails 6.0 doesn't have the necessary plumbing to allow us to
massage the BSON value into something that can be interpreted as
numeric. Rails 6.0 is EOL, though, and rather than copy-and-paste
 the entire validate_each implementation to support a dead version
of Rails, we'll just skip the test when that version of Rails
is installed.

* Work around MongoDB 8.3.0-alpha database validation

* Can't assume DB will be autocreated with MongoDB 8.3.0-alpha

---------

Co-authored-by: Jamis Buck <jamis.buck@mongodb.com>
Co-authored-by: Dmitry Rybakov <dmitry.rybakov@mongodb.com>